+++ packages/iov-bns/types/bnsconnection.d.ts
@@ -37,19 +37,19 @@ export declare class BnsConnection implements BcpAtomicSwapConnection {
     /**
      * All matching swaps that are open (from app state)
      */
-    getSwapFromState(query: BcpSwapQuery): Promise<ReadonlyArray<BcpAtomicSwap>>;
+    getSwapsFromState(query: BcpSwapQuery): Promise<ReadonlyArray<BcpAtomicSwap>>;
     /**
      * All matching swaps that are open (in app state)
      *
      * To get claimed and returned, we need to look at the transactions.... TODO
      */
-    getSwap(query: BcpSwapQuery): Promise<ReadonlyArray<BcpAtomicSwap>>;
+    getSwaps(query: BcpSwapQuery): Promise<ReadonlyArray<BcpAtomicSwap>>;
     /**
      * Emits currentState (getSwap) as a stream, then sends updates for any matching swap
      *
      * This includes an open swap beind claimed/expired as well as a new matching swap being offered
      */
-    watchSwap(query: BcpSwapQuery): Stream<BcpAtomicSwap>;
+    watchSwaps(query: BcpSwapQuery): Stream<BcpAtomicSwap>;
     searchTx(txQuery: BcpTxQuery): Promise<ReadonlyArray<ConfirmedTransaction | FailedTransaction>>;
     /**
      * A stream of all transactions that match the tags from the present moment on